update to most recent diplomat
