+++ ffi/diplomat/Cargo.toml
@@ -65,8 +65,8 @@ writeable = { version = "0.4", path = "../../utils/writeable/" }
 # Run `cargo make diplomat-install` to get the right diplomat binary installed
 # The version here can either be a `version = ".."` spec or `git = "https://github.com/rust-diplomat/diplomat", rev = ".."`
 # Since this crate is published, Diplomat must be published preceding a new ICU4X release but may use git versions in between
-diplomat = { git = "https://github.com/rust-diplomat/diplomat", rev = "9bae290384548bd2a81babcced6608064079c3d9" }
-diplomat-runtime = { git = "https://github.com/rust-diplomat/diplomat", rev = "9bae290384548bd2a81babcced6608064079c3d9" }
+diplomat = { git = "https://github.com/rust-diplomat/diplomat", rev = "e80b5969a77f9f49e751f63ff51f2af230027c96" }
+diplomat-runtime = { git = "https://github.com/rust-diplomat/diplomat", rev = "e80b5969a77f9f49e751f63ff51f2af230027c96" }
 icu_testdata = { version = "0.6", path = "../../provider/testdata", optional = true }
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
