+++ ffi/diplomat/src/wasm_glue.rs
@@ -3,37 +3,8 @@
 // (online at: https://github.com/unicode-org/icu4x/blob/main/LICENSE ).
 
 use alloc::format;
-use rust_std::ffi::CString;
-use rust_std::io;
-use rust_std::os::raw::c_char;
-
+use diplomat_runtime::{console_log, console_warn};
 use log::{Level, LevelFilter, Metadata, Record};
-
-// minimal WASM logger based on https://github.com/DeMille/wasm-glue
-extern "C" {
-    fn log_js(ptr: *const c_char);
-    fn warn_js(ptr: *const c_char);
-}
-
-fn _log(buf: &str) -> io::Result<()> {
-    let cstring = CString::new(buf)?;
-
-    unsafe {
-        log_js(cstring.as_ptr());
-    }
-
-    Ok(())
-}
-
-fn _warn(buf: &str) -> io::Result<()> {
-    let cstring = CString::new(buf)?;
-
-    unsafe {
-        warn_js(cstring.as_ptr());
-    }
-
-    Ok(())
-}
 
 struct ConsoleLogger;
 
@@ -44,9 +15,9 @@ impl log::Log for ConsoleLogger {
 
     fn log(&self, record: &Record) {
         if record.level() <= Level::Warn {
-            _warn(format!("[{}] {}", record.level(), record.args()).as_str()).unwrap();
+            console_warn(format!("[{}] {}", record.level(), record.args()).as_str()).unwrap();
         } else {
-            _log(format!("[{}] {}", record.level(), record.args()).as_str()).unwrap();
+            console_log(format!("[{}] {}", record.level(), record.args()).as_str()).unwrap();
         }
     }
