+++ ffi/diplomat/wasm/.gitignore
@@ -1,0 +1,1 @@
+lib/*.wasm
