+++ ffi/diplomat/wasm/diplomat.config.js
@@ -1,0 +1,4 @@
+export default {
+    wasm_path: new URL('./lib/icu_capi_cdylib.wasm', import.meta.url),
+    init: wasm => wasm.icu4x_init(),
+};
