+++ ffi/diplomat/wasm/example/main.mjs
@@ -1,13 +1,13 @@
-import { ICU4XFixedDecimal, ICU4XDataProvider, ICU4XLocale, ICU4XFixedDecimalFormatter, ICU4XFixedDecimalGroupingStrategy } from "../lib/api.mjs"
+import { ICU4XFixedDecimal, ICU4XDataProvider, ICU4XLocale, ICU4XFixedDecimalFormatter, ICU4XFixedDecimalGroupingStrategy } from "../lib/index.js"
 
 const decimal = ICU4XFixedDecimal.create(1234);
 decimal.multiply_pow10(-2);
 decimal.set_sign("Negative");
 console.log(decimal.to_string());
 
-const dataProvider = ICU4XDataProvider.create_test().provider;
+const dataProvider = ICU4XDataProvider.create_test();
 
 const locale = ICU4XLocale.create("bn");
 
-const format = ICU4XFixedDecimalFormatter.try_new(locale, dataProvider, "Auto").fdf;
-console.log(format.format(decimal));
+const fdf = ICU4XFixedDecimalFormatter.try_new(locale, dataProvider, "Auto");
+console.log(fdf.format(decimal));
